dev: move authentication to proxy middleware
